moved the controller functions from the routes into the controllers folder
